requirements.txt
Flask==3.0.0
itsdangerous==2.1.2
nltk==3.8.1
numpy==1.24.4
python-decouple==3.3
six
sqlparse==0.3.0
textblob==0.15.3
tzlocal==2.0.0
Werkzeug==3.0.0
scikit-learn==1.3.2
pandas==1.3.5
python-dotenv